+++ packages/manager/src/constants.ts
@@ -314,3 +314,9 @@ export const MOCK_SERVICE_WORKER =
 
 // Maximum payment methods
 export const MAXIMUM_PAYMENT_METHODS = 6;
+
+// Price of LKE's High Availability offering in USD
+export const HIGH_AVAILABILITY_PRICE =
+  process.env.REACT_APP_LKE_HIGH_AVAILABILITY_PRICE === undefined
+    ? undefined
+    : Number(process.env.REACT_APP_LKE_HIGH_AVAILABILITY_PRICE);