+++ packages/manager/src/features/Volumes/VolumeTableRow.tsx
@@ -1,19 +1,26 @@
-import { Event } from '@linode/api-v4/lib/account';
+import {
+  Event,
+  EventAction,
+  NotificationType,
+} from '@linode/api-v4/lib/account';
 import * as React from 'react';
-import { Link, useLocation } from 'react-router-dom';
+import { Link, useHistory, useLocation } from 'react-router-dom';
 import { compose } from 'recompose';
+import Chip from 'src/components/core/Chip';
 import Hidden from 'src/components/core/Hidden';
-import { makeStyles } from 'src/components/core/styles';
+import { makeStyles, Theme } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
 import Grid from 'src/components/Grid';
 import LinearProgress from 'src/components/LinearProgress';
 import TableCell from 'src/components/TableCell';
 import TableRow from 'src/components/TableRow';
+import useNotifications from 'src/hooks/useNotifications';
 import { formatRegion } from 'src/utilities';
 import { ExtendedVolume } from './types';
 import VolumesActionMenu, { ActionHandlers } from './VolumesActionMenu';
+import useEvents from 'src/hooks/useEvents';
 
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles((theme: Theme) => ({
   volumePath: {
     width: '35%',
     wordBreak: 'break-all',
@@ -28,6 +35,33 @@ const useStyles = makeStyles(() => ({
       We can remove once we make the full switch to CMR styling
       */
     paddingRight: '0 !important',
+  },
+  chipWrapper: {
+    alignSelf: 'center',
+  },
+  chip: {
+    borderRadius: 1,
+    fontSize: '0.65rem',
+    marginTop: 0,
+    marginBottom: 0,
+    marginLeft: theme.spacing(2),
+    minHeight: theme.spacing(2),
+    paddingLeft: theme.spacing(0.5),
+    paddingRight: theme.spacing(0.5),
+  },
+  forceUpgradeChip: {
+    backgroundColor: theme.color.chipButton,
+    '&:hover': {
+      backgroundColor: theme.color.chipButtonHover,
+    },
+  },
+  upgradePendingChip: {
+    backgroundColor: 'transparent',
+    border: '1px solid #ccc',
+  },
+  nvmeChip: {
+    backgroundColor: 'transparent',
+    border: '1px solid #02B159',
   },
 }));
 
@@ -62,16 +96,42 @@ export const VolumeTableRow: React.FC<CombinedProps> = (props) => {
     size,
     recentEvent,
     region,
+    hardware_type: hardwareType,
     filesystem_path: filesystemPath,
     linodeLabel,
     linode_id: linodeId,
     linodeStatus,
   } = props;
 
+  const { events } = useEvents();
+  const history = useHistory();
   const location = useLocation();
+  const notifications = useNotifications();
   const isVolumesLanding = Boolean(location.pathname.match(/volumes/));
 
   const formattedRegion = formatRegion(region);
+
+  const isNVMe = hardwareType === 'nvme';
+
+  const eligibleForUpgradeToNVMe = notifications.some(
+    (notification) =>
+      notification.type ===
+        ('volume_migration_scheduled' as NotificationType) &&
+      notification.entity?.id === id
+  );
+
+  const nvmeUpgradeScheduledByUserImminent = notifications.some(
+    (notification) =>
+      notification.type === ('volume_migration_imminent' as NotificationType) &&
+      notification.entity?.id === id
+  );
+
+  const nvmeUpgradeScheduledByUserInProgress = events.some(
+    (event) =>
+      event.action === ('volume_migrate' as EventAction) &&
+      event.entity?.id === id &&
+      event.status === 'started'
+  );
 
   return isUpdating ? (
     <TableRow
@@ -93,19 +153,68 @@ export const VolumeTableRow: React.FC<CombinedProps> = (props) => {
   ) : (
     <TableRow key={`volume-row-${id}`} data-qa-volume-cell={id}>
       <TableCell data-qa-volume-cell-label={label}>
-        <Grid container wrap="nowrap" alignItems="center">
-          <Grid item>
-            <div>{label}</div>
-          </Grid>
+        <Grid
+          container
+          wrap="nowrap"
+          justify="space-between"
+          alignItems="flex-end"
+        >
+          {isVolumesLanding ? (
+            <>
+              <Grid item>
+                <div>{label}</div>
+              </Grid>
+              {isNVMe ? (
+                <Grid item className={classes.chipWrapper}>
+                  <Chip
+                    className={`${classes.chip} ${classes.nvmeChip}`}
+                    label="NVMe"
+                    data-testid="nvme-chip"
+                  />
+                </Grid>
+              ) : linodeId &&
+                eligibleForUpgradeToNVMe &&
+                !nvmeUpgradeScheduledByUserImminent ? (
+                <Grid item className={classes.chipWrapper}>
+                  <Chip
+                    className={`${classes.chip} ${classes.forceUpgradeChip}`}
+                    label="UPGRADE TO NVMe"
+                    onClick={() => history.push(`/linodes/${linodeId}/upgrade`)}
+                    data-testid="upgrade-chip"
+                  />
+                </Grid>
+              ) : linodeId &&
+                (nvmeUpgradeScheduledByUserImminent ||
+                  nvmeUpgradeScheduledByUserInProgress) ? (
+                <Grid item className={classes.chipWrapper}>
+                  <Chip
+                    className={`${classes.chip} ${classes.upgradePendingChip}`}
+                    label="UPGRADE PENDING"
+                    data-testid="upgrading-chip"
+                  />
+                </Grid>
+              ) : null}
+            </>
+          ) : (
+            <Grid item>
+              <div>{label}</div>
+            </Grid>
+          )}
         </Grid>
       </TableCell>
-      {region && <TableCell data-qa-volume-region>{formattedRegion}</TableCell>}
-      <TableCell data-qa-volume-size>{size} GB</TableCell>
-      <Hidden xsDown>
-        <TableCell className={classes.volumePath} data-qa-fs-path>
-          {filesystemPath}
+      {region ? (
+        <TableCell data-qa-volume-region noWrap>
+          {formattedRegion}
         </TableCell>
-      </Hidden>
+      ) : null}
+      <TableCell data-qa-volume-size>{size} GB</TableCell>
+      {!isVolumesLanding ? (
+        <Hidden xsDown>
+          <TableCell className={classes.volumePath} data-qa-fs-path>
+            {filesystemPath}
+          </TableCell>
+        </Hidden>
+      ) : null}
       {isVolumesLanding && (
         <TableCell data-qa-volume-cell-attachment={linodeLabel}>
           {linodeId ? (