+++ packages/api-v4/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [2021-11-09] - v0.62.0
+### Added:
+- Types, methods, and interfaces related to Block Storage migrations
+
+## [2021-11-01] - v0.61.0
+### Added:
+- Support for enabling Kubernetes High Availability
+- Endpoints for getting Kubernetes Dashboard and resetting Kube config
+
 ## [2021-10-20] - v0.60.0
 
 ### Changed: